# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  phone?: string;
  clinicName: string;
  gstNumber?: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
  role: string;
  clinicName: string;
  gstNumber?: string;
  clinicAddress?: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface UpdateSettingsRequest {
  name?: string;
  phone?: string;
  clinicName?: string;
  gstNumber?: string;
  clinicAddress?: string;
}

export interface Patient {
  id: number;
  name: string;
  phone: string;
  email?: string;
  address?: string;
  age?: number;
  gender?: string;
  medicalHistory?: string;
  notes?: string;
  firstVisit?: string;
  createdAt?: string;
}

export interface Appointment {
  id: number;
  patientId: number;
  patientName?: string;
  doctor: string;
  date: string;
  time: string;
  treatment?: string;
  notes?: string;
  status: string;
  createdAt?: string;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  patientId: number;
  patientName?: string;
  date: string;
  subtotal: number;
  discount?: number;
  gst?: number;
  total: number;
  paymentMethod?: string;
  paymentStatus: string;
  createdAt?: string;
}

export type PatientDetail = Patient & {
  appointments?: Appointment[];
  invoices?: Invoice[];
};

export interface CreatePatientRequest {
  name: string;
  phone: string;
  email?: string;
  address?: string;
  age?: number;
  gender?: string;
  medicalHistory?: string;
  notes?: string;
  firstVisit?: string;
}

export interface CreateAppointmentRequest {
  patientId: number;
  doctor: string;
  date: string;
  time: string;
  treatment?: string;
  notes?: string;
  status?: string;
}

export interface UpdateAppointmentRequest {
  patientId?: number;
  doctor?: string;
  date?: string;
  time?: string;
  treatment?: string;
  notes?: string;
  status?: string;
}

export interface Treatment {
  id: number;
  name: string;
  cost: number;
  duration?: number;
  description?: string;
}

export interface CreateTreatmentRequest {
  name: string;
  cost: number;
  duration?: number;
  description?: string;
}

export interface InvoiceItem {
  id: number;
  invoiceId: number;
  treatment: string;
  price: number;
}

export type InvoiceDetail = Invoice & {
  items?: InvoiceItem[];
};

export interface CreateInvoiceItem {
  treatment: string;
  price: number;
}

export interface CreateInvoiceRequest {
  patientId: number;
  date?: string;
  discount?: number;
  paymentMethod?: string;
  paymentStatus?: string;
  items: CreateInvoiceItem[];
}

export interface UpdateInvoiceRequest {
  paymentMethod?: string;
  paymentStatus?: string;
  discount?: number;
}

export interface Staff {
  id: number;
  name: string;
  role: string;
  phone?: string;
  email?: string;
  specialization?: string;
  joiningDate?: string;
  salary?: number;
  status?: string;
}

export interface CreateStaffRequest {
  name: string;
  role: string;
  phone?: string;
  email?: string;
  specialization?: string;
  joiningDate?: string;
  salary?: number;
  status?: string;
}

export interface DashboardStats {
  todayAppointments: number;
  totalPatients: number;
  monthlyRevenue: number;
  pendingPayments: number;
  completedTreatments: number;
  newPatients: number;
}

export interface ChartDataPoint {
  label: string;
  value: number;
}

export interface RevenueReport {
  totalRevenue: number;
  totalInvoices: number;
  paidAmount: number;
  pendingAmount: number;
  invoices?: Invoice[];
}

export interface TreatmentStat {
  treatment: string;
  count: number;
  revenue: number;
}

export type ListPatientsParams = {
  search?: string;
};

export type ListAppointmentsParams = {
  date?: string;
  status?: string;
};

export type ListInvoicesParams = {
  status?: string;
};

export type GetDailyRevenueParams = {
  date?: string;
};

export type GetMonthlyRevenueParams = {
  month?: string;
  year?: string;
};