# lib/api-zod/src/generated/types/invoiceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from "./invoice";
import type { InvoiceItem } from "./invoiceItem";

export type InvoiceDetail = Invoice & {
  items?: InvoiceItem[];
};