# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  phone: string;
  email?: string;
  address?: string;
  age?: number;
  gender?: string;
  medicalHistory?: string;
  notes?: string;
  firstVisit?: string;
  createdAt?: string;
}