# lib/api-zod/src/generated/types/updateInvoiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateInvoiceRequest {
  paymentMethod?: string;
  paymentStatus?: string;
  discount?: number;
}